requirements.txt
Jinja2
PyYAML
bs4
django-htmlmin
lxml
peewee
requests
